+++ src/elements/button-link/src/index.tsx
@@ -2,7 +2,6 @@
 
 import * as React from 'react'
 import { jsx, Styled, useThemeUI } from 'theme-ui'
-import { Glyph, Icon } from '@uswitch/trustyle.icon'
 import get from '@uswitch/trustyle-utils.get'
 
 type Overwrite<T, U> = Omit<T, keyof U> & U
@@ -26,8 +25,6 @@ export const ButtonLink = <
   children,
   variant,
   size = 'large',
-  beforeIcon,
-  afterIcon,
   ...props
 }: Props<T>) => {
   const { theme }: any = useThemeUI()
@@ -64,21 +61,7 @@ export const ButtonLink = <
       }}
       {...props}
     >
-      {beforeIcon && (
-        <span
-          sx={{ variant: `elements.buttons.variants.${variant}.beforeIcon` }}
-        >
-          <Icon color="white" glyph={beforeIcon as Glyph} direction="left" />
-        </span>
-      )}
       {children}
-      {afterIcon && (
-        <span
-          sx={{ variant: `elements.buttons.variants.${variant}.afterIcon` }}
-        >
-          <Icon color="white" glyph={afterIcon as Glyph} direction="right" />
-        </span>
-      )}
     </Styled.a>
   )
 }