+++ src/themes/uswitch/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.30.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.uswitch-theme@2.30.2...@uswitch/trustyle.uswitch-theme@2.30.3) (2020-11-11)
+
+
+### Bug Fixes
+
+* adds black friday variant to badges. fixes visual issues with hero. ([a81f239](https://github.com/uswitch/trustyle/commit/a81f239))
+
+
+
+
+
 ## [2.30.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.uswitch-theme@2.30.1...@uswitch/trustyle.uswitch-theme@2.30.2) (2020-11-11)
 
 **Note:** Version bump only for package @uswitch/trustyle.uswitch-theme